Fix flannel.service to survive reboot.
The flannel.service attempted to create a key
'/coreos.com/network/config' for its network, and would fail to start
if this key already existed.

Since this is an example, it's probably safest to blow away the
key before attempting to (re)create; this way, if the user has
somehow assigned a bad value, the assumptions in the rest of the config
will still be valid.