Add Calico network plugin for Vagrant provisioner.
Adds support for Calico networking in Kubernetes. Nodes are now
joined in a BGP mesh. When a pod is created, a new route for that
pod is distributed to all other Kubernetes nodes. This means
that network traffic between all containers on the network is
un-encapsulated (no need for an overlay network).

Calico also provides ACL function, which may be configured using
the calicoctl tool (or programmed directly into the etcd
datastore). This allows intent-based firewalling at both the
source and destination pod.