Fix kubernetes 1.1 binary link
Previous link was broken (either incorrect or the Github release was deleted). Use a Google-hosted URL for now.

Note that the canonical location will be `https://github.com/kubernetes/kubernetes/releases/tag/v1.1.1`, but that is currently a pre-release and doesn't have the binary; this is a stopgap fix until the Github release is released.